requirements.txt
google-genai
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
pillow
psycopg
psycopg2-binary